# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Monitor API",
        "description": "API documentation for Uptime Monito project",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/badge/{monitorId}": {
            "get": {
                "tags": [
                    "Badge"
                ],
                "summary": "Get badge for a monitor",
                "description": "Returns an HTML badge representing the status of a monitor.",
                "operationId": "getBadge",
                "parameters": [
                    {
                        "name": "monitorId",
                        "in": "path",
                        "description": "ID of the monitor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HTML badge",
                        "content": {
                            "text/html": {
                                "example": "<img alt='Static Badge' src='https://img.shields.io/badge/badge-label-up-green'>"
                            }
                        }
                    },
                    "404": {
                        "description": "Monitor not found"
                    }
                }
            }
        },
        "/api/v1/monitors": {
            "get": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Get list of monitors",
                "operationId": "7011373c0b11938e685b61c3f0c6ab6c",
                "responses": {
                    "200": {
                        "description": "List of monitors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Monitor"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Create a new monitor",
                "operationId": "c63bd78e77bb794346ccd2513ebb0ce3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Monitor"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Monitor created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Monitor"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/monitors/{monitor}": {
            "get": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Get monitor detail",
                "operationId": "ff7181d511cc5aeb61120488ca8d49cc",
                "parameters": [
                    {
                        "name": "monitor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Monitor detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Monitor"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Update monitor",
                "operationId": "4f22e2e007581a9589cf92c26268bf92",
                "parameters": [
                    {
                        "name": "monitor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Monitor"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Monitor updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Monitor"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Delete monitor",
                "operationId": "323e9abff49e7c220dd0cf011f4ec14e",
                "parameters": [
                    {
                        "name": "monitor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Monitor deleted"
                    }
                }
            }
        },
        "/api/v1/monitors/{monitor}/history": {
            "get": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Get monitor history",
                "operationId": "d007756cc3e1f8bb02d482f0bc1721d4",
                "parameters": [
                    {
                        "name": "monitor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "mode",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "status",
                                "response",
                                "latency"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Monitor history",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/monitors/{monitor}/calendar-summary": {
            "get": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Get monitor calendar summary",
                "operationId": "4325ec3a36c64e8bb4e7ca8b2813fab9",
                "parameters": [
                    {
                        "name": "monitor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calendar summary",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Monitor": {
                "title": "Monitor",
                "required": [
                    "project_id",
                    "label",
                    "periodicity",
                    "type",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "project_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "label": {
                        "type": "string",
                        "example": "Ping Google"
                    },
                    "periodicity": {
                        "description": "Interval in seconds (5-300)",
                        "type": "integer",
                        "maximum": 300,
                        "minimum": 5,
                        "example": 60
                    },
                    "monitor_type": {
                        "type": "string",
                        "example": "ping"
                    },
                    "badge_label": {
                        "type": "string",
                        "example": "Google"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "latest_status": {
                        "type": "string",
                        "example": "succeeded"
                    },
                    "hostname": {
                        "type": "string",
                        "example": "google.com"
                    },
                    "port": {
                        "type": "integer",
                        "example": 80
                    },
                    "url": {
                        "type": "string",
                        "example": "https://google.com"
                    },
                    "check_status": {
                        "type": "boolean",
                        "example": true
                    },
                    "keywords": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}